#1 create recomend plants widget
