+++ lib/screens/home/components/body.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:plant_app/constants.dart';
-// import 'package:plant_app/constants.dart';
-import 'package:plant_app/screens/home/components/title_with_more_btn.dart';
-// import 'package:flutter_svg/svg.dart';
-// import 'package:plant_app/constants.dart';
 
 import 'header_with_searchbox.dart';
+import 'title_with_more_btn.dart';
+import 'recomend_plants.dart';
+
+// import 'package:plant_app/constants.dart';
+// import 'package:flutter_svg/svg.dart';
+// import 'package:plant_app/constants.dart';
 
 class Body extends StatelessWidget {
   const Body({super.key});
@@ -24,94 +26,7 @@ class Body extends StatelessWidget {
           TitleWithMoreBtn(title: 'Recomended', press: () {}),
           // it will cover 40% of our total width
           // 全幅の40％をカバーする
-          RecomendPlantCard(
-            image: 'assets/images/image_1.png',
-            title: 'Samantha',
-            country: 'Russia',
-            price: 440,
-            press: () {},
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class RecomendPlantCard extends StatelessWidget {
-  const RecomendPlantCard({
-    super.key,
-    required this.image,
-    required this.title,
-    required this.country,
-    required this.price,
-    required this.press,
-  });
-
-  final String image, title, country;
-  final int price;
-  final void Function()? press;
-
-  @override
-  Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
-    return Container(
-      margin: const EdgeInsets.only(
-        left: kDefaultPadding,
-        top: kDefaultPadding / 2,
-        bottom: kDefaultPadding * 2.5,
-      ),
-      width: size.width * 0.4,
-      child: Column(
-        children: [
-          Image.asset(image),
-          GestureDetector(
-            onTap: press,
-            child: Container(
-              padding: const EdgeInsets.all(kDefaultPadding / 2),
-              decoration: BoxDecoration(
-                color: Colors.white,
-                borderRadius: const BorderRadius.only(
-                  bottomLeft: Radius.circular(10),
-                  bottomRight: Radius.circular(10),
-                ),
-                boxShadow: [
-                  BoxShadow(
-                    offset: const Offset(0, 10),
-                    blurRadius: 50,
-                    color: kPrimaryColor.withOpacity(0.23),
-                  ),
-                ],
-              ),
-              child: Row(
-                children: [
-                  RichText(
-                    text: TextSpan(
-                      children: [
-                        TextSpan(
-                          text: '$title\n'.toUpperCase(),
-                          style: Theme.of(context).textTheme.labelLarge,
-                        ),
-                        TextSpan(
-                          text: country.toUpperCase(),
-                          style: TextStyle(
-                            color: kPrimaryColor.withOpacity(0.5),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  const Spacer(),
-                  Text(
-                    '\$$price',
-                    style: Theme.of(context)
-                        .textTheme
-                        .labelLarge
-                        ?.copyWith(color: kPrimaryColor),
-                  ),
-                ],
-              ),
-            ),
-          )
+          const RecomendsPlants(),
         ],
       ),
     );
